.
Species populations is severely limited in this iteration. Stable
populations can and do appear, but they do not dominate the world.

It seems that there is a 'dominant' species: 25-40 members, a constantly
growing average generation number (600+). This species doesn't kill off
everyone though. The second species was a somewhat equal population, but
a younger generation count.

There seems to be fluxuations in population sizes. The 'dominant'
species sometimes falls down to dangerously low levels (18-25) while
another species takes over. However the other species doesn't seem to
stabalize. The previously dominant species seems to be resilient to
significant population changes. And it eventually regains control.

As I was typing the previous paragraph, the exact situation happened
again: The dominant species fell down into the teens (extremely close to
extinction), while the 'second' species took over. But then the second
species was exterminated entirely (not sure if it was killed off, or if
it died off). This led to the first species resuming control over the
environment.