Explosive growth
Purposely increased food production and consumption and rate of
mutation. This allows multiple species to maintain high and stable
populations.

With increased population stability, species can spread and
differentiate more than before. There was a green species that split
into two separate species (while still living side-by-side).

Species also manage to spread to new areas of the map and still survive
(because food isn't as scarce). This increases the survivability of the
species because one subsection of the population may die off, but the
other will survive.

Within minutes, species manage to move around the map and invade other
locations of the map.

----

With 5 species, 4 of them maintained a stable population (with gen
numbers above 1k). The fifth species hasn't managed to get passed 20
creatures (with stability), whereas the other 4 creatures are in the
range of 50-100.