Visualizing brains
Added the ability to click on a creature and see some of its internal
structure. However, the brains are too large to easily visualize. Not
sure how useful this feature is.

Reformatted the structure of the brains in order to allow for greater
scalability. The previous brains were hard-coded with this structure:
inputs->hidden->outputs. And the number of hidden layers was fixed. I
expaned the brains to allow for any number of hidden layers. Although
creatures do not yet possess the ability to grow the number of hidden
layers (or the number of neurons per layer).

Unfortunately (or fortunately?), there was a slight change in some of
the math which has led to slightly different behavior.
The threshold for movement used to be 0.01, but with the changed brains,
that threshold was never reached among the starting creatures (due to
arbitrary constants). This led to creatures never exploring. I lowed the
threshold to 0.001 and noticed that movement resumed. Then I lowered it
to 0.0001 and noticed that there was still movement.

Therefore, this threshold seems to have an impact on the population, but
it is hard to identifiy the exact response. A lower threshold might
increase coin-flippy behavior, or it might allow for more fine-tuned
motion. It's hard to tell.