.
Creatures can stabalize, and consistently there are species that are a
couple thousand generations old. But eventually these species go
extinct.

At any one time there seems to be at least 2 species that are 'old'
while the other 3 are fighting for survival, but getting nowhere.

The creatures seem to evolve well, even when food is quite sparce.
Perhaps they are resorting to eating each other more often, but it is
hard to tell.