Complete Overhaul
Environments can now be created, and different biomes can be described
in order to create a custom feel (and to test evolution approaches).

--------

I wanted to add more complex environments (with the ability to
distinguish between an infinite number of biomes).

Within seconds of running the first working copy (post overhaul),
creatures spread from South America into North America, and then Asia,
and then Africa/Australia. I wondered how long it would take for the
creatures to evolve to use different environments, but I found that they
still exploded easily.

There was a purple species that exploded and took over the map. But then
a green species caused a great disruption. The green species was either
a very sudden mutation from the purple, or it had been living there for
so long but I never saw it. What seems particulary interesting is that
it is very hard to see the green species because it is living with green
territory. I wonder if it was a coincidence that the very-obvious-purple
species was wiped out due to their color.

Many species seem to go through a population boom and then there is a
period of peace, adn then the population collapses. This has occured so
frequently that I am beginning to think it is a bug.

I just witnessed a dark purple explode and spread to many corners of the
map (short lived). Then a green species did something similar. Then a
mutation of the original purple, now a light purple, spread quite far.