Complexity
With the new biome parameters, it is significantly harder to find
balance.

----

Experimented with adding an additional layer to the creatures brains.
The resulting populations were significantly higher and signficantly
more stable.

With the higher level thinking, it appears that the creatures have
decided to not eat themselves. This allows the creatures to live
significantly closer to each other, and therefore use the evironment
more efficiently.